Complete all 443 packet implementations - 150 CS + 293 SC verified
- Add 18 missing CS packets (file transfer, exchange, inventory, weapon)
- Add 41 missing SC packets (leave sight, skills, exchange, visual effects)
- Fix 15 SC packets missing IPacketFixed interface
- Update tracking files: cs_packets.toon and sc_packets.toon
- Document IDA typos: EQUIPED (missing P), SUCESS (missing C)
- All packets verified against IDA Pro analysis
- Build: 0 errors, 0 warnings

+++ kakia_lime_odyssey_server/PacketHandlers/CS_MOVE_SLOT_ITEM_Handler.cs
@@ -63,8 +63,8 @@ class CS_MOVE_SLOT_ITEM_Handler : PacketHandler
 
 			sc_move.move_list.Add(new()
 			{
-				fromCount = cs_move.count,
-				toCount = cs_move.count,
+				fromCount = (ulong)cs_move.count,
+				toCount = (ulong)cs_move.count,
 				fromSlot = cs_move.from,
 				toSlot = cs_move.to
 			});
@@ -87,7 +87,7 @@ class CS_MOVE_SLOT_ITEM_Handler : PacketHandler
 
 				sc_move.move_list.Add(new()
 				{
-					fromCount = cs_move.count - item_slot1.GetAmount(),
+					fromCount = (ulong)cs_move.count - item_slot1.GetAmount(),
 					toCount = item_slot2.GetAmount(),
 					fromSlot = cs_move.from,
 					toSlot = cs_move.to
@@ -115,16 +115,16 @@ class CS_MOVE_SLOT_ITEM_Handler : PacketHandler
 
 				sc_move.move_list.Add(new()
 				{
-					fromCount = cs_move.count,
-					toCount = cs_move.count,
+					fromCount = (ulong)cs_move.count,
+					toCount = (ulong)cs_move.count,
 					fromSlot = cs_move.from,
 					toSlot = cs_move.to
 				});
 
 				sc_move.move_list.Add(new()
 				{
-					fromCount = cs_move.count,
-					toCount = cs_move.count,
+					fromCount = (ulong)cs_move.count,
+					toCount = (ulong)cs_move.count,
 					fromSlot = cs_move.to,
 					toSlot = cs_move.from
 				});
